Return show files sorted by location

+++ tests/util/test_common.py
@@ -16,7 +16,7 @@ from autosubliminal.util.common import get_today, run_cmd, connect_url, wait_for
     to_list, to_obj_or_list, to_dict, get_boolean, safe_text, safe_lowercase, safe_uppercase, safe_trim, sanitize, \
     display_mapping_dict, display_list_single_line, display_list_multi_line, display_value, display_item_title, \
     display_item_name, display_interval, display_timestamp, convert_timestamp, humanize_bytes, get_common_path, \
-    get_root_path, get_file_size, set_rw_and_remove
+    get_root_path, get_file_size, set_rw_and_remove, atoi, natural_keys
 
 vcr = VCR(path_transformer=VCR.ensure_suffix('.yaml'),
           record_mode='once',
@@ -379,6 +379,7 @@ def test_get_root_path():
 
 
 def test_get_file_size():
+    file_path = None
     try:
         fd, file_path = tempfile.mkstemp(text=True)
         file = open(file_path, 'w')
@@ -388,7 +389,8 @@ def test_get_file_size():
         size = get_file_size(file_path).split(' ')[0]
         assert float(size) > 0
     finally:
-        os.remove(file_path)
+        if os.path.exists(file_path):
+            os.remove(file_path)
 
 
 def test_get_file_size_exception():
@@ -396,6 +398,7 @@ def test_get_file_size_exception():
 
 
 def test_set_rw_and_remove():
+    file_path = None
     try:
         fd, file_path = tempfile.mkstemp(text=True)
         file = open(file_path, 'w')
@@ -407,3 +410,16 @@ def test_set_rw_and_remove():
     finally:
         if os.path.exists(file_path):
             os.remove(file_path)
+
+
+def test_atoi():
+    assert atoi('test') == 'test'
+    assert atoi('test01') == 'test01'
+    assert atoi('01') == 1
+    assert atoi('1') == 1
+
+
+def test_natural_keys():
+    my_list = ['season 23', 'season 15', 'season 30', 'season 05', 'season 01', 'root']
+    my_sorted_list = ['root', 'season 01', 'season 05', 'season 15', 'season 23', 'season 30']
+    assert sorted(my_list, key=natural_keys) == my_sorted_list
